==> R/4.1-ts_model_CA.R <==
#####################################
########## Load packages ############
#####################################
library(forecast)
library(reshape2)
library(ggplot2)
library(mgcv)
library(TSA)
library(dplyr)
library(fda)
library(sp)
library(gstat)
library(fields)
library(MBA)
library(Vizumap)
library(nlme)
library(autoimage)
library(FRK)
library(pwr)
library(MTS)
library(tseries)
library(RColorBrewer)
library(dglm)
library(quantreg)
setwd("~/Dropbox/UCSD/Thesis/3.Precipitation/Code")
source("varx_fixed.R")

## get aspect ratio from lon/lat
map_aspect = function(x, y) {
  x.center <- sum(range(x)) / 2
  y.center <- sum(range(y)) / 2
  x.dist <- ggplot2:::dist_central_angle(x.center + c(-0.5, 0.5), rep(y.center, 2))
  y.dist <- ggplot2:::dist_central_angle(rep(x.center, 2), y.center + c(-0.5, 0.5))
  y.dist / x.dist
}

durbinWatsonTest.r <- function(resid, max.lag=1){
  n <-  length(resid)
  dw <- rep(0, max.lag)
  den <- sum(resid^2)
  for (lag in 1:max.lag){
    dw[lag] <- (sum((resid[(lag+1):n] - resid[1:(n-lag)])^2))/den
  }
  dw
}

durbinWatsonTest.glm <- function(model, max.lag=1, simulate=TRUE, reps=1000, 
                                method="resample", 
                                alternative="two.sided"){
  resid <- residuals(model)
  if (any(is.na(resid))) stop ('residuals include missing values')
  n <- length(resid)
  r <- dw <-rep(0, max.lag)
  den <- sum(resid^2)
  for (lag in 1:max.lag){
    dw[lag] <- (sum((resid[(lag+1):n] - resid[1:(n-lag)])^2))/den
    r[lag] <- (sum(resid[(lag+1):n]*resid[1:(n-lag)]))/den
  }
  if (!simulate){
    result <- list(r=r, dw=dw)
    result
  }
  else {
    X <- model.matrix(model)
    mu <- fitted.values(model)
    Y <-  matrix(sample(resid, n*reps, replace=TRUE), n, reps) + matrix(mu, n, reps)
    E <- apply(Y, 2, function(y) residuals(glm(y[y>0] ~ X[y>0,], family=Gamma(link = "log"))))
    if(is.list(E))
    {
      DW = unlist(lapply(E, function(x) durbinWatsonTest.r(x,max.lag=max.lag)))
    }else
      DW <- apply(E, 2, durbinWatsonTest.r, max.lag=max.lag)
    if (max.lag == 1) DW <- rbind(DW)
    p <- rep(0, max.lag)
    if (alternative == 'two.sided'){
      for (lag in 1:max.lag) {
        p[lag] <- (sum(dw[lag] < DW[lag,]))/reps
        p[lag] <- 2*(min(p[lag], 1 - p[lag]))
      }
    }
    else if (alternative == 'positive'){
      for (lag in 1:max.lag) {
        p[lag] <- (sum(dw[lag] > DW[lag,]))/reps
      }
    }
    else {
      for (lag in 1:max.lag) {
        p[lag] <- (sum(dw[lag] < DW[lag,]))/reps
      }
    }
    result <- list(r=r, dw=dw, p=p, alternative=alternative)
    result
  }
}

#####################################
############## Load data ############
#####################################
tas_path = "/Users/wenyilin/Documents/R/NA-CORDEX/data/rds/tas-rec-rcp85-mon-44i/"
pr_path = "/Users/wenyilin/Documents/R/NA-CORDEX/data/rds/pr-rec-rcp85-mon-44i/"
map_path = "/Users/wenyilin/Documents/R/NA-CORDEX/map/"
code_path = "/Users/wenyilin/Dropbox/UCSD/Thesis/3.Precipitation/Code/"
res_path = "/Users/wenyilin/Dropbox/UCSD/Thesis/3.Precipitation/Code/results/"
within_ca = readRDS(file = paste0(map_path,"within_ca.rds"))
load(paste0(map_path,"within_rec_ca.rdata"))
load(paste0(map_path,"ca_elevation_rec.rdata"))
load("/Users/wenyilin/Dropbox/UCSD/Thesis/3.Precipitation/Code/reports/ca_geodata.rdata")

### list files in CA
#coords_ca_all = list.files(path = data_path,pattern = "^coords_ca")
tas_hist_ca_all = list.files(path = tas_path,pattern = "^tas_ca_tas.hist")
tas_rcp_ca_all = list.files(path = tas_path,pattern = "^tas_ca_tas.rcp85")
pr_hist_ca_all = list.files(path = pr_path,pattern = "^pr_ca_pr.hist")
pr_rcp_ca_all = list.files(path = pr_path,pattern = "^pr_ca_pr.rcp85")

### example analysis for CA
tas_hist_ca = readRDS(file = paste0(tas_path,tas_hist_ca_all[1]))
tas_rcp_ca = readRDS(file = paste0(tas_path,tas_rcp_ca_all[1]))
pr_hist_ca = readRDS(file = paste0(pr_path,pr_hist_ca_all[1]))
pr_rcp_ca = readRDS(file = paste0(pr_path,pr_rcp_ca_all[1]))


## find non-ocean area
coords_ca$ind = 1
tmp.coords = merge(ca_elevation, coords_ca, by = c("latitude","longitude"),all.x=TRUE)
land.ind = which(tmp.coords$ind==1 | tmp.coords$elevation_geonames.x>0)
ca_elevation = ca_elevation[land.ind,]
tas_hist_ca = tas_hist_ca[land.ind,]
pr_hist_ca = pr_hist_ca[land.ind,]
ca_elevation$ele_std = (ca_elevation$elevation_geonames - min(ca_elevation$elevation_geonames))/(max(ca_elevation$elevation_geonames)-min(ca_elevation$elevation_geonames))
ca_elevation$ele_norm = (ca_elevation$elevation_geonames - mean(ca_elevation$elevation_geonames))/1000

#####################################
####### Seasonal Aggregation ########
#####################################

## temperature
time_dat = data.frame(year = rep(1950:2005,each=12),
                     month = rep(1:12,56),
                     ts.point = 1:(12*56))  ## historical measures

# time_dat = data.frame(year = rep(2006:2100,each=12),
#                      month = rep(1:12,95),
#                      ts.point = 1:(12*95))  ## future projection

gls.dat = data.frame(lon = ca_elevation[,1], lat = ca_elevation[,2],
                     elevation = ca_elevation[,3],
                     tas = tas_hist_ca,
                     loc = as.factor(1:length(ca_elevation[,1])))
gls.dat.long = reshape2::melt(gls.dat,id.vars = c("lon","lat","elevation","loc"))
names(gls.dat.long) = c("lon","lat","elevation","loc","ts.point","tas")
gls.dat.long$ts.point = as.numeric(gsub("\\D", "", gls.dat.long$ts.point))
gls.dat.long = merge(gls.dat.long,time_dat, by.x = "ts.point",all.x = TRUE)
gls.dat.summary = as.data.frame(gls.dat.long %>% 
                                  filter(month %in% c(1,2,3)) %>%
                                  group_by(lon,lat,elevation,loc,year) %>%
                                  dplyr::summarise(winter = mean(tas))
)
gls.dat.summary$spring = as.data.frame(gls.dat.long %>% 
                                         filter(month %in% c(4,5,6)) %>%
                                         group_by(lon,lat,elevation,loc,year) %>%
                                         dplyr::summarise(spring = mean(tas))
)$spring

gls.dat.summary$summer = as.data.frame(gls.dat.long %>% 
                                         filter(month %in% c(7,8,9)) %>%
                                         group_by(lon,lat,elevation,loc,year) %>%
                                         dplyr::summarise(summer = mean(tas))
)$summer

gls.dat.summary$fall = as.data.frame(gls.dat.long %>% 
                                       filter(month %in% c(10,11,12)) %>%
                                       group_by(lon,lat,elevation,loc,year) %>%
                                       dplyr::summarise(fall = mean(tas))
)$fall
gls.tas.summary = gls.dat.summary[order(gls.dat.summary$loc,gls.dat.summary$year),]

## Precipitation
gls.dat = data.frame(lon = ca_elevation[,1], lat = ca_elevation[,2],
                     elevation = ca_elevation[,3],
                     tas = pr_hist_ca*86400*30,
                     loc = as.factor(1:length(ca_elevation[,1])))
gls.dat.long = reshape2::melt(gls.dat,id.vars = c("lon","lat","elevation","loc"))
names(gls.dat.long) = c("lon","lat","elevation","loc","ts.point","tas")
gls.dat.long$ts.point = as.numeric(gsub("\\D", "", gls.dat.long$ts.point))
gls.dat.long = merge(gls.dat.long,time_dat, by.x = "ts.point",all.x = TRUE)
gls.dat.summary = as.data.frame(gls.dat.long %>% 
                                  filter(month %in% c(1,2,3)) %>%
                                  group_by(lon,lat,elevation,loc,year) %>%
                                  dplyr::summarise(winter = mean(tas))
)
gls.dat.summary$spring = as.data.frame(gls.dat.long %>% 
                                         filter(month %in% c(4,5,6)) %>%
                                         group_by(lon,lat,elevation,loc,year) %>%
                                         dplyr::summarise(spring = mean(tas))
)$spring

gls.dat.summary$summer = as.data.frame(gls.dat.long %>% 
                                         filter(month %in% c(7,8,9)) %>%
                                         group_by(lon,lat,elevation,loc,year) %>%
                                         dplyr::summarise(summer = mean(tas))
)$summer

gls.dat.summary$fall = as.data.frame(gls.dat.long %>% 
                                       filter(month %in% c(10,11,12)) %>%
                                       group_by(lon,lat,elevation,loc,year) %>%
                                       dplyr::summarise(fall = mean(tas))
)$fall
gls.pr.summary = gls.dat.summary[order(gls.dat.summary$loc,gls.dat.summary$year),]

#####################################
###########  plotting ###############
#####################################

season.var = c("winter","spring","summer","fall" )
loc.names = c("SF","SD","Yosemite","Death Valley")
select.loc = data.frame(sf = c(37.75,-122.25), sd = c(32.75,-117.25),
                        yo = c(37.75,-119.25), dv = c(36.25,-116.75))
select.ind = c(126, 1, 132, 87)
select.loc.col = c("red","green","blue","darkgrey")
season.var = c("winter","spring","summer","fall")
sub.loc = list(x = c(-122.25, -117.25, -119.25, -116.75), 
               y = c(37.75,32.75,37.75,36.25),
               labels = loc.names)
n.season = length(season.var)
n.loc = length(unique(gls.pr.summary$loc))
time.yr = (range(gls.pr.summary$year)[1] : range(gls.pr.summary$year)[2])
#time.rcp.yr = (range(gls.pr.rcp$year)[1] : range(gls.pr.rcp$year)[2])
#time.pts = c(time.yr - mean(time.yr))/sd(time.yr)
time.pts = c(time.yr - mean(time.yr))/10
#time.rcp.pts = c(time.rcp.yr - mean(time.rcp.yr))/10
n.time = length(time.pts)
#n.rcp.time = length(time.rcp.pts)

gls.tas.summary$time.pts = time.pts
#gls.tas.rcp$time.pts = time.rcp.pts
#gls.tas.all = rbind(gls.tas.summary,gls.tas.rcp)

gls.pr.summary$time.pts = time.pts
#gls.pr.rcp$time.pts = time.rcp.pts
#gls.pr.all = rbind(gls.pr.summary,gls.pr.rcp)

#png("figures/tas_4season_map.png",width = 250, height = 100, res = 100,units = "in")
par(oma = c(0, 0, 3, 0), mar = c(2, 1, 1, 1))
autoimage(ca_elevation$longitude,ca_elevation$latitude,
          gls.tas.summary[gls.tas.summary$year==1992,season.var],
          interp.args = list(no.X = 200, no.Y = 200),
          map = "county",ylab = "Latitude",xlab = "Longitude",
          main = season.var,
          #main.args = list(cex = 2),
          outer.title = "Average seasonal temperature (C) in 1992",
          mtext.args = list(cex = 1),
          #text = as,
          #text.args = list(cex=2),
          legend.axis.args = list(cex.axis=1),
          #breaks = seq(-5,30,by=0.01),
          #axis.args = list(at = seq(-5,30,by=5),labels = seq(-5,30,by=5)),
          #legend.axis.args = list(at = seq(-5,30,by=5),labels = seq(-5,30,by=5), las = 2),
          zlim = c(-6,31),
          #col=rev(cmap),
          size = c(1, 4), lratio = 0.35,
          legend = "vertical")
#dev.off()

cmap <- rev(colorspace::sequential_hcl(n = 100, palette = "Viridis"))
par(oma = c(0, 0, 3, 0), mar = c(2, 1, 1, 1))
autoimage(coords_ca$longitude,coords_ca$latitude,
          gls.pr.summary[gls.pr.summary$year==1992,season.var],
          interp.args = list(no.X = 200, no.Y = 200),
          map = "county",ylab = "Latitude",xlab = "Longitude",
          main = season.var,
          outer.title = "Average seasonal precipitation (mm/month) in 1992",
          mtext.args = list(cex = 1),
          #text = as,
          #text.args = list(cex=2),
          legend.axis.args = list(cex.axis=1),
          #breaks = seq(-300,0,by=0.1),
          #axis.args = list(at = seq(300,0,length=8),labels = seq(100,0,length=7)),
          #legend.axis.args = list(at = seq(-300,0,by=50),labels = seq(300,0,by=-50), las = 2),
          zlim = c(0,200),
          col=cmap,
          size = c(1, 4), lratio = 0.35,
          legend = "vertical")
#legend.scale(c(-10,35), horizontal = FALSE)

#### Display in time series
season.var = c("winter","spring","summer","fall" )
loc.names = c("SF","SD","Yosemite","Death Valley")
n.season = length(season.var)
n.loc = length(unique(gls.tas.summary$loc))
time.yr = (range(gls.tas.summary$year)[1] : range(gls.tas.summary$year)[2])
time.pts = c(time.yr - mean(time.yr))/sd(time.yr)
n.time = length(time.pts)

par(mfrow = c(2,2), oma=c(1, 1, 2,5),  mar = c(2, 2, 2, 2))
for(s in 1:length(select.ind))
{
  tas.dat = ts(gls.tas.summary[gls.tas.summary$loc==select.ind[s],season.var[1]],start = 1950)
  plot(tas.dat, ylim = range(gls.tas.summary[,season.var]), main = loc.names[s],
       xlab = "Year", ylab = "Average temperature")
  for(i in 2: n.season)
  {
    tas.dat = ts(gls.tas.summary[gls.tas.summary$loc==select.ind[s],season.var[i]],start = 1950)
    lines(tas.dat, ylim = range(gls.tas.summary[,season.var]),col=i)
  }
  #if(s==1){
  #  legend("topright",bty='n', xpd=NA,cex = 0.8,
  #       legend=var.names,lty = 1, col=1:length(var.names))
  #}
  #{legend(par('usr')[2], par('usr')[4],bty='n', xpd=NA,cex = 0.8,
  #       legend=var.names,lty = 1, col=1:length(var.names))}
}
legend(par('usr')[2], par('usr')[4]*2, bty='n', xpd=NA,cex = 1.1,
       legend=season.var,lty = 1, col=1:n.season)
mtext("Average seasonal temperature (C) at four typical locations", outer = TRUE, cex = 1.1)

par(mfrow = c(2,2), oma=c(1, 1, 2,5),  mar = c(2, 2, 2, 2))
for(s in 1:length(select.ind))
{
  tas.dat = ts(gls.pr.summary[gls.pr.summary$loc==select.ind[s],season.var[1]],start = 1950)
  plot(tas.dat, ylim = c(0,300), main = loc.names[s],
       xlab = "Year", ylab = "Average precipitation (mm/month)")
  for(i in 2: length(season.var))
  {
    tas.dat = ts(gls.pr.summary[gls.pr.summary$loc==select.ind[s],season.var[i]],start = 1950)
    lines(tas.dat,col=i)
  }
  #if(s==1){
  #  legend("topright",bty='n', xpd=NA,cex = 0.8,
  #         legend=var.names,lty = 1, col=1:length(var.names))
  #}
  #{legend(par('usr')[2], par('usr')[4],bty='n', xpd=NA,cex = 0.8,
  #       legend=var.names,lty = 1, col=1:length(var.names))}
}
legend(par('usr')[2], par('usr')[4]*2, bty='n', xpd=NA,cex = 1.1,
       legend=season.var,lty = 1, col=1:n.season)
mtext("Average seasonal precipitation (mm/month) at four typical locations", outer = TRUE, cex = 1.1)

#####################################
######### TS model selection ########
#####################################

# Temperature
## 1. normal test with shapiro test
plot_list = function(x, main = "", xlab = ""){
  maxy = sapply(x, function(j) j$y)
  plot(x[[1]], ylim = range(maxy), xlab = xlab, main = main,# xlim = c(-3,3),
       col = "darkgray", lwd = 0.5)
  for (j in seq_along(x)) {
    lines(x[[j]], col = "darkgray", lwd = 0.5)
  }
}

norm_test = matrix(0, nrow = n.loc, ncol = n.season)

for(j in 1:n.season)
{
  d.resids = list()
  for(i in 1:n.loc)
  {
    tas_exp = gls.tas.summary[gls.tas.summary$loc==i,season.var[j]]
    lm.fit = lm(tas_exp ~ time.pts)
    d.resids[[i]] = density(residuals(lm.fit))
    norm_test[i,j] = shapiro.test(residuals(lm.fit))$p.value
  }
  plot_list(d.resids, main = season.var[j])
}

normal.res = table(norm_test < 0.05)
names(normal.res) = c("normal","non-normal")
normal.res

## 2. ACF and PACF
lm_acf = lm_pacf = matrix(NA, nrow = n.loc, ncol = n.season)
resid_mat = array(0, dim = c(n.loc, n.season, n.time))
acf_sig = qnorm(1 - 0.05 / 2) / sqrt(n.time)
box_pval = matrix(0, nrow = n.loc, ncol = n.season)

for(i in 1:n.loc)
{
  for(j in 1:n.season)
  {
    tas_exp = gls.tas.summary[gls.tas.summary$loc==i,season.var[j]]
    #tas_out = tas_exp[-1]
    #tas_lag1 = tas_exp[-n.time]
    lm.fit = lm(tas_exp ~ time.pts)
    resid_mat[i,j,] = residuals(lm.fit)
    acf.res = acf(residuals(lm.fit),plot = FALSE)
    lm_acf[i,j] = acf.res$acf[1,,]
    pacf.res = pacf(residuals(lm.fit),plot = FALSE)
    lm_pacf[i,j] = pacf.res$acf[1,,]
    box_pval[i,j] = Box.test(resid_mat[i,j,],type = "Ljung-Box",lag = 10)$p.value
  }
}

boxplot(lm_pacf, ylim = c(-acf_sig-0.1,acf_sig+0.1))
abline(h = acf_sig,col="red",lty=2)
abline(h = -acf_sig,col="red",lty=2)

x = box_pval[order(box_pval)]
plot(ecdf(box_pval))
abline(0,1)
lines(x, x +2*sqrt(x*(1-x)/90))
lines(x, x -2*sqrt(x*(1-x)/90))

## 3. cross-correlation between seasons
multi.coef = multi.pval = matrix(0, nrow = n.loc, ncol = n.season)
for(i in 1:n.loc)
{
  tas_exp = gls.tas.summary[gls.tas.summary$loc==i,season.var]
  # Step 1: seasonal correlation
  season.sub.m1 = lm(tas_exp$spring ~ tas_exp$winter)
  season.sub.m2 = lm(tas_exp$summer ~ tas_exp$spring)
  season.sub.m3 = lm(tas_exp$fall ~ tas_exp$summer)
  season.sub.m4 = lm(tas_exp$winter[2:n.time] ~ tas_exp$fall[1:(n.time-1)])
  multi.coef[i,] = round(c(coef(season.sub.m1)[2],coef(season.sub.m2)[2],
                           coef(season.sub.m3)[2],coef(season.sub.m4)[2]),4)
  multi.pval[i,] = round(c(summary(season.sub.m1)$coefficients[2,4],
                           summary(season.sub.m2)$coefficients[2,4],
                           summary(season.sub.m3)$coefficients[2,4],
                           summary(season.sub.m4)$coefficients[2,4]),4)
}
colnames(multi.coef) = colnames(multi.pval) = c("sp ~ wi","su ~ sp","fa ~ su","wi ~ fa")
boxplot(multi.pval)
abline(h=0.05,col="red")

## 4. Test with AIC
aic_all = matrix(0,nrow = n.loc, ncol = 6)
for(i in 1:n.loc)
{
  tas_exp = gls.tas.summary[gls.tas.summary$loc==i,season.var]
  m_ar1 = varx_fixed(tas_exp,p=1,xt = matrix(time.pts, nrow = length(time.pts)))
  fixed_beta = rbind(rep(1,4),
                     diag(1,4),
                     rep(1,4))
  m_ar1_1 = varx_fixed(tas_exp[-1,],p=1,
                       xt = matrix(time.pts, nrow = length(time.pts)),
                       fixed = fixed_beta)
  m_ar0 = varx_fixed(tas_exp,p=0,xt = matrix(time.pts, nrow = length(time.pts)))
  fixed_beta = rbind(matrix(1,nrow = 6,ncol = 4),
                     diag(1,4))
  m_ar1_cor = varx_fixed(tas_exp[-1,],p=1,xt = x_multi,fixed = fixed_beta)
  fixed_beta = rbind(rep(1,4),
                     diag(1,4),
                     rep(1,4),
                     diag(1,4))
  m_ar1_nocor = varx_fixed(tas_exp[-1,],p=1,xt = x_multi,fixed = fixed_beta)
  fixed_beta = rbind(matrix(1,nrow = 2,ncol = 4),diag(1,4))
  m_ar0_cor = varx_fixed(tas_exp[-1,],p=0,xt = x_multi,fixed = fixed_beta)
  aic_all[i,] = round(c(m_ar1$aic,m_ar1_1$aic,m_ar0$aic,m_ar1_cor$aic,m_ar1_nocor$aic,m_ar0_cor$aic),4)
}
colnames(aic_all) = c("AR1 + t","nocorAR1 + t","t", "AR1 + t + season_lag","nocorAR1 + t +season_lag","t+season_lag")

## precipitation

### 1. residual autocorrelation
lm_acf = lm_pacf = matrix(NA, nrow = n.loc, ncol = n.season)
dw_p = matrix(NA, nrow = n.loc, ncol = n.season)
resid_mat = array(0, dim = c(n.loc, n.season, n.time))
acf_sig = qnorm(1 - 0.05 / 2) / sqrt(n.time)

for(i in 1:n.loc)
{
  for(j in 1:n.season)
  {
    tas_exp = gls.pr.summary[gls.pr.summary$loc==i,season.var[j]]
    glm.fit = glm(tas_exp ~ time.pts, family=Gamma(link = "log"))
    resid_mat[i,j,] = residuals(glm.fit, type="deviance")
    dw_p[i,j] = durbinWatsonTest.glm(glm.fit)$p
    acf.res = acf(residuals(glm.fit),plot = FALSE)
    lm_acf[i,j] = acf.res$acf[1,,]
    pacf.res = pacf(residuals(glm.fit),plot = FALSE)
    lm_pacf[i,j] = pacf.res$acf[1,,]
  }
}

boxplot(lm_pacf, ylim = c(-acf_sig-0.1,acf_sig+0.1))
abline(h = acf_sig,col="red",lty=2)
abline(h = -acf_sig,col="red",lty=2)

## 2. cross-seasonal effect
multi.coef = multi.pval = matrix(0, nrow = n.loc, ncol = n.season)
for(i in 1:n.loc)
{
  tas_exp = gls.pr.summary[gls.pr.summary$loc==i,season.var]
  # Step 1: seasonal correlation
  season.sub.m1 = lm(tas_exp$spring ~ tas_exp$winter)
  season.sub.m2 = lm(tas_exp$summer ~ tas_exp$spring)
  season.sub.m3 = lm(tas_exp$fall ~ tas_exp$summer)
  season.sub.m4 = lm(tas_exp$winter[2:n.time] ~ tas_exp$fall[1:(n.time-1)])
  multi.coef[i,] = round(c(coef(season.sub.m1)[2],coef(season.sub.m2)[2],
                           coef(season.sub.m3)[2],coef(season.sub.m4)[2]),4)
  multi.pval[i,] = round(c(summary(season.sub.m1)$coefficients[2,4],
                           summary(season.sub.m2)$coefficients[2,4],
                           summary(season.sub.m3)$coefficients[2,4],
                           summary(season.sub.m4)$coefficients[2,4]),4)
}
colnames(multi.coef) = colnames(multi.pval) = c("sp ~ wi","su ~ sp","fa ~ su","wi ~ fa")
boxplot(multi.pval)
abline(h=0.05,col="red")

#####################################
######### TS model fitting ##########
#####################################

### temperature
center_apply <- function(x) {
  apply(x, 2, function(y) y - mean(y))
}
beta_est = beta_se = matrix(0, nrow = n.loc, ncol = n.season)
season_coef = season_se = matrix(0, nrow = n.loc, ncol = n.season)
sresids = resids = array(NA, dim = c(n.loc, n.time-1, n.season))
for(i in 1:n.loc)
{
  tas_exp = gls.tas.summary[gls.tas.summary$loc==i,season.var]
  x_multi = cbind(time.pts[-1],tas_exp$fall[1:(n.time-1)],
                  tas_exp$winter[2:n.time],tas_exp$spring[2:n.time],
                  tas_exp$summer[2:n.time])
  fixed_beta2 = rbind(matrix(1,nrow = 2,ncol = 4),
                      c(1,0,0,0),
                      c(0,1,0,0),
                      c(0,0,1,1),
                      c(0,0,0,0))
  m_ar0_cor2 = varx_fixed(tas_exp[-1,],p=0,xt = x_multi,fixed = fixed_beta2)
  beta_est[i,] = m_ar0_cor2$beta[,1]
  beta_se[i,] = m_ar0_cor2$se.beta[,1]
  season_coef[i,] = c(m_ar0_cor2$beta[1,2],m_ar0_cor2$beta[2,3],
                      m_ar0_cor2$beta[3,4],
                      m_ar0_cor2$beta[4,4])
  season_se[i,] = c(m_ar0_cor2$se.beta[1,2],m_ar0_cor2$se.beta[2,3],
                    m_ar0_cor2$se.beta[3,4],
                    m_ar0_cor2$se.beta[4,4])
  resids[i,,] = m_ar0_cor2$residuals
  #resids_c = center_apply(m_ar0_cor2$residuals)
  sresids[i,,] = t((chol(solve(m_ar0_cor2$Sigma)))%*% t(m_ar0_cor2$residuals))
}
cmap <- colorRampPalette(rev(brewer.pal(11, "RdBu")))

par(oma = c(0, 0, 3, 0), mar = c(2, 1, 1, 1))
autoimage(coords_ca$longitude,coords_ca$latitude,
          beta_est,
          # interp.args = list(no.X = 200, no.Y = 200),
          map = "county",ylab = "Latitude",xlab = "Longitude",
          main = season.var,
          outer.title = "Estimated slopes of four seasons",
          mtext.args = list(cex = 1),
          #text = as,
          #text.args = list(cex=1),
          legend.axis.args = list(cex.axis=1),
          size = c(1, 4), lratio = 0.35,
          col=cmap(100),
          zlim = c(-1,1),
          legend = "vertical")

z_score = beta_est/beta_se
par(oma = c(0, 0, 3, 0), mar = c(2, 1, 1, 1))
autoimage(coords_ca$longitude,coords_ca$latitude,
          z_score,
          # interp.args = list(no.X = 200, no.Y = 200),
          map = "county",ylab = "Latitude",xlab = "Longitude",
          main = season.var,
          outer.title = "Z_scores of estimated slopes of four seasons",
          mtext.args = list(cex = 1),
          #text = as,
          #text.args = list(cex=1),
          legend.axis.args = list(cex.axis=1),
          size = c(1,4), lratio = 0.35,
          col=cmap(100),
          zlim = c(-5.5,5.5),
          legend = "vertical")

### precipitation
###### model checking

lm_acf = lm_pacf = matrix(NA, nrow = n.loc, ncol = n.season)
dw_p = matrix(NA, nrow = n.loc, ncol = n.season)
resid_mat = array(0, dim = c(n.loc, n.season, n.time))
acf_sig = qnorm(1 - 0.05 / 2) / sqrt(n.time)

for(i in 1:n.loc)
{
  for(j in 1:n.season)
  {
    tas_exp = gls.pr.summary[gls.pr.summary$loc==i,season.var[j]]
    glm.fit = glm(tas_exp ~ time.pts, family=Gamma(link = "log"))
    resid_mat[i,j,] = residuals(lm.fit)
    dw_p[i,j] = durbinWatsonTest.glm(glm.fit)$p
    acf.res = acf(residuals(lm.fit),plot = FALSE)
    lm_acf[i,j] = acf.res$acf[1,,]
    pacf.res = pacf(residuals(lm.fit),plot = FALSE)
    lm_pacf[i,j] = pacf.res$acf[1,,]
  }
}

boxplot(lm_pacf, ylim = c(-acf_sig-0.1,acf_sig+0.1))
abline(h = acf_sig,col="red",lty=2)
abline(h = -acf_sig,col="red",lty=2)

multi.coef = multi.pval = matrix(0, nrow = n.loc, ncol = n.season)
for(i in 1:n.loc)
{
  tas_exp = gls.pr.summary[gls.pr.summary$loc==i,season.var]
  # Step 1: seasonal correlation
  season.sub.m1 = lm(tas_exp$spring ~ tas_exp$winter)
  season.sub.m2 = lm(tas_exp$summer ~ tas_exp$spring)
  season.sub.m3 = lm(tas_exp$fall ~ tas_exp$summer)
  season.sub.m4 = lm(tas_exp$winter[2:n.time] ~ tas_exp$fall[1:(n.time-1)])
  multi.coef[i,] = round(c(coef(season.sub.m1)[2],coef(season.sub.m2)[2],
                           coef(season.sub.m3)[2],coef(season.sub.m4)[2]),4)
  multi.pval[i,] = round(c(summary(season.sub.m1)$coefficients[2,4],
                           summary(season.sub.m2)$coefficients[2,4],
                           summary(season.sub.m3)$coefficients[2,4],
                           summary(season.sub.m4)$coefficients[2,4]),4)
}
colnames(multi.coef) = colnames(multi.pval) = c("sp ~ wi","su ~ sp","fa ~ su","wi ~ fa")
boxplot(multi.pval)
abline(h=0.05,col="red")

slope_est = slope_se =  p.value = gof.test = matrix(0, nrow = n.loc, ncol = n.season)
sig.level = 0.1
for(i in 1:n.loc)
{
  for(j in 1:n.season)
  {
    tas_exp = gls.pr.summary[gls.pr.summary$loc==i,season.var[j]]
    fit1 <- glm(tas_exp ~ time.pts, family=Gamma(link = "log"))
    deviance = fit1$deviance
    p.value[i,j] = pchisq(deviance, df = fit1$df.residual, lower.tail = F)
    gof.test[i,j] = ifelse(p.value[i,j] < sig.level,0,1)
    shape_est <- 1/summary(fit1)$dispersion
    scale_est <- 1/fit1$coefficients[1]/shape_est
    slope_est[i,j] <- fit1$coefficients[2]
    slope_se[i,j] <- summary(fit1)$coefficients[2,2]
  }
}

cmap <- colorRampPalette(brewer.pal(11, "RdBu"))

par(oma = c(0, 0, 3, 0), mar = c(2, 1, 1, 1))
autoimage(coords_ca$longitude,coords_ca$latitude,
          slope_est,
          #interp.args = list(no.X = 200, no.Y = 200),
          map = "county",ylab = "Latitude",xlab = "Longitude",
          main = season.var,
          outer.title = "Estimated slopes (gamma regression) of four seasons",
          mtext.args = list(cex = 1),
          #text = as,
          #text.args = list(cex=1),
          legend.axis.args = list(cex.axis=1),
          size = c(1,4), lratio = 0.35,
          col=cmap(100),
          zlim = c(-0.4,0.4),
          legend = "vertical")

z_score = slope_est/slope_se
par(oma = c(0, 0, 3, 0), mar = c(2, 1, 1, 1))
autoimage(coords_ca$longitude,coords_ca$latitude,
          z_score,
          # interp.args = list(no.X = 200, no.Y = 200),
          map = "county",ylab = "Latitude",xlab = "Longitude",
          main = season.var,
          outer.title = "Z-scores of estimated slopes (gamma regression)",
          mtext.args = list(cex = 1),
          #text = as,
          #text.args = list(cex=1),
          legend.axis.args = list(cex.axis=1),
          size = c(1,4), lratio = 0.35,
          col=cmap(100),
          zlim = c(-5,5),
          legend = "vertical")
